refactor: 여행기 저장 및 공유 API 수정 (PR 피드백2차) #37
